milvus/requirements.txt
pymilvus>=2.3.0
pillow
torch
torchvision
transformers
matplotlib
tqdm
numpy
